Lesson 150.
Added code in FeedVC to show messages in reverse order (i.e. newest first).
Added a function to get userEmail from userID in DataService, and implemented this function in cellForRowAt indexPath in FeedVC.
Added code in signOutBtnWasPressed in MeVC to allow user to sign out; on signOut the loginVC will be instantiated.
Added a check in AuthVC to see if user is logged in. If logged in, then the AuthVC is dismissed.
Updated emailLbl in CreatePost and MeVCs to show current user’s e-mail instead of boilerplate e-mail.